Improvements to build system and additional documentation

+++ tools/package.sh
@@ -12,19 +12,27 @@
 #
 # Run the Image
 #
-# $ docker run --rm -v `pwd`:/io:Z ezralanglois/interop sh /io/tools/package.sh /io /io/dist travis OFF Release
-# $ docker run --rm -v `pwd`:/io:Z ezralanglois/interop sh /io/tools/package.sh /io /io/dist teamcity OFF Release
+# $ docker run --rm -w /tmp --user `id -u`:`id -g` -v `pwd`:/src:ro -v `pwd`/dist:/dist:rw ezralanglois/interop sh /src/tools/package.sh /src /dist travis OFF Release
+# $ docker run --rm -w /tmp --user `id -u`:`id -g` -v `pwd`:/src:ro -v `pwd`/dist:/dist:rw ezralanglois/interop sh /src/tools/package.sh /src /dist teamcity OFF Release
 #
 # Debug the Image Interactively
 #
-# $ docker run --it -v `pwd`:/io ezralanglois/interop sh /io/tools/package_linux.sh /io /io/dist
+# $ docker run --rm -i -t -v `pwd`:/io ezralanglois/interop sh /io/tools/package_linux.sh /io /io/dist
 #
 #
 ########################################################################################################################
-set -e
+set -ex
 INTEROP_C89=OFF
 BUILD_TYPE=Release
-THREAD_COUNT=4
+
+# When inside docker and not using root, but using the root home. So, change it to working directory
+# Without this, nuget fails when trying to read Nuget.config
+whoami 1>/dev/null 2>&1 || export HOME=$PWD
+
+# Get value from environment for low memory vms
+if [ -z $THREAD_COUNT ] ; then
+    THREAD_COUNT=4
+fi
 
 if [ ! -z $1 ] ; then
     SOURCE_PATH=$1
@@ -40,7 +48,7 @@ fi
 if [ ! -z $3 ] ; then
     BUILD_SERVER=$3
     DISABLE_SUBDIR=OFF
-    if [[ "$server" == "travis" ]]; then
+    if [[ "$BUILD_SERVER" == "travis" ]]; then
         DISABLE_SUBDIR=ON
     fi
 else
@@ -88,29 +96,32 @@ fi
 mkdir $BUILD_PATH
 
 if [ -e $ARTIFACT_PATH ] ; then
-    rm -fr $ARTIFACT_PATH
+    rm -fr $ARTIFACT_PATH/*
+else
+    mkdir $ARTIFACT_PATH
 fi
-mkdir $ARTIFACT_PATH
 
 
 # Build Python Wheels for a range of Python Versions
-for PYBUILD in `ls -1 /opt/python`; do
-    PYTHON_BIN=/opt/python/${PYBUILD}/bin
-    if [[ "$PYBUILD" == cp26* ]]; then
-        continue
-    fi
-    if [[ "$PYBUILD" == cp33* ]]; then
-        continue
-    fi
-    touch $SOURCE_PATH/src/ext/python/CMakeLists.txt
-    run "Configure ${PYBUILD}" cmake $SOURCE_PATH -B${BUILD_PATH} -DPYTHON_EXECUTABLE=${PYTHON_BIN}/python ${CMAKE_EXTRA_FLAGS} -DSKIP_PACKAGE_ALL_WHEEL=ON -DPYTHON_WHEEL_PREFIX=${ARTIFACT_PATH}/tmp
+if [ -z $PYTHON_VERSION ] && [  -e /opt/python ] ; then
+    for PYBUILD in `ls -1 /opt/python`; do
+        PYTHON_BIN=/opt/python/${PYBUILD}/bin
+        if [[ "$PYBUILD" == cp26* ]]; then
+            continue
+        fi
+        if [[ "$PYBUILD" == cp33* ]]; then
+            continue
+        fi
+        rm -fr ${BUILD_PATH}/src/ext/python/*
+        run "Configure ${PYBUILD}" cmake $SOURCE_PATH -B${BUILD_PATH} -DPYTHON_EXECUTABLE=${PYTHON_BIN}/python ${CMAKE_EXTRA_FLAGS} -DSKIP_PACKAGE_ALL_WHEEL=ON -DPYTHON_WHEEL_PREFIX=${ARTIFACT_PATH}/tmp
 
-    run "Test ${PYBUILD}" cmake --build $BUILD_PATH --target check -- -j${THREAD_COUNT}
-    run "Build ${PYBUILD}" cmake --build $BUILD_PATH --target package_wheel -- -j${THREAD_COUNT}
-    auditwheel show ${ARTIFACT_PATH}/tmp/interop*${PYBUILD}*linux_x86_64.whl
-    auditwheel repair ${ARTIFACT_PATH}/tmp/interop*${PYBUILD}*linux_x86_64.whl -w ${ARTIFACT_PATH}
-    rm -fr ${ARTIFACT_PATH}/tmp
-done
+        run "Test ${PYBUILD}" cmake --build $BUILD_PATH --target check -- -j${THREAD_COUNT}
+        run "Build ${PYBUILD}" cmake --build $BUILD_PATH --target package_wheel -- -j${THREAD_COUNT}
+        auditwheel show ${ARTIFACT_PATH}/tmp/interop*${PYBUILD}*linux_x86_64.whl
+        auditwheel repair ${ARTIFACT_PATH}/tmp/interop*${PYBUILD}*linux_x86_64.whl -w ${ARTIFACT_PATH}
+        rm -fr ${ARTIFACT_PATH}/tmp
+    done
+fi
 
 if [ ! -z $PYTHON_VERSION ] ; then
     if [ "$PYTHON_VERSION" == "ALL" ] ; then
@@ -122,17 +133,23 @@ if [ ! -z $PYTHON_VERSION ] ; then
         echo "Building Python $py_ver"
         if hash pyenv 2> /dev/null; then
             export PATH=$(pyenv root)/shims:${PATH}
-            pyenv install $py_ver || true
+            if [[ "$OSTYPE" == "linux-gnu" ]]; then
+                if hash patchelf 2> /dev/null; then
+                    pyenv install 3.5 -s  || true
+                    pyenv global 3.5 || true
+                    pip3 install auditwheel==1.5
+                fi
+            fi
+            pyenv install $py_ver -s  || true
             pyenv global $py_ver || true
+            if [[ "$OSTYPE" == "darwin"* ]]; then
+                pip install delocate
+            fi
             which python
+            which pip
             python --version
             pip install numpy
             pip install wheel
-            if [[ "$OSTYPE" == "linux-gnu" ]]; then
-                pip install auditwheel==1.5
-            elif [[ "$OSTYPE" == "darwin"* ]]; then
-                pip install delocate
-            fi
         fi
         run "Configure $py_ver" cmake $SOURCE_PATH -B${BUILD_PATH} ${CMAKE_EXTRA_FLAGS} -DENABLE_PYTHON_DYNAMIC_LOAD=ON -DPYTHON_EXECUTABLE=`which python` -DPYTHON_WHEEL_PREFIX=${ARTIFACT_PATH}/tmp
         run "Build $py_ver" cmake --build $BUILD_PATH -- -j${THREAD_COUNT}
@@ -144,10 +161,11 @@ if [ ! -z $PYTHON_VERSION ] ; then
             delocate-wheel $ARTIFACT_PATH/tmp/*.whl
             delocate-addplat -c --rm-orig -x 10_9 -x 10_10 $ARTIFACT_PATH/tmp/*.whl -w $ARTIFACT_PATH
         elif hash auditwheel 2> /dev/null; then
+            auditwheel --version
             auditwheel show ${ARTIFACT_PATH}/tmp/*.whl
             auditwheel repair ${ARTIFACT_PATH}/tmp/*.whl -w ${ARTIFACT_PATH}
         fi
-        touch $SOURCE_PATH/src/ext/python/CMakeLists.txt
+        rm -fr ${BUILD_PATH}/src/ext/python/*
         rm -fr ${ARTIFACT_PATH}/tmp
     done
 fi
@@ -161,7 +179,7 @@ fi
 run "Package" cmake --build $BUILD_PATH --target bundle -- -j${THREAD_COUNT}
 
 if hash dotnet 2> /dev/null; then
-    run "Configure DotNetCore" cmake $SOURCE_PATH -B${BUILD_PATH} ${CMAKE_EXTRA_FLAGS} -DCSBUILD_TOOL=DotNetCore && cmake --$BUILD_PATH --target nupack -- -j${THREAD_COUNT} || true
+    run "Configure DotNetCore" cmake $SOURCE_PATH -B${BUILD_PATH} ${CMAKE_EXTRA_FLAGS} -DCSBUILD_TOOL=DotNetCore && cmake --build $BUILD_PATH --target nupack -- -j${THREAD_COUNT} || true
 fi
 rm -fr ${ARTIFACT_PATH}/tmp
 echo "List Artifacts:"
